+++ packages/datadog_flutter_plugin/integration_test_app/integration_test/rum_manual_error_reporting_test.dart
@@ -70,6 +70,7 @@ void main() {
     var manualError = view.errorEvents[1];
     expect(manualError.message, contains('Rum error message'));
     expect(manualError.source, kIsWeb ? 'custom' : 'network');
+    expect(manualError.fingerprint, 'custom-fingerprint');
 
     var thrownError = view.errorEvents[2];
     expect(thrownError.message, contains('This was an error!'));